# api/openapi.yaml
openapi: 3.0.3
info:
  title: Creatorr API
  version: 0.1.0
  description: |
    REST API for Creatorr (creator VOD daemon).
    HTMX HTML routes and SSE (`/api/events`) are out of this contract - see AGENTS.md.
servers:
  - url: /
    description: Local / Docker
...
paths:
...
  /api/tasks:
    get:
      tags: [tasks]
      operationId: listTasks
      summary: Running and pending tasks by domain queue
      responses:
        "200":
          description: Active tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      tags: [tasks]
      operationId: enqueueTask
      summary: Enqueue a task (manual / tests)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnqueueTaskRequest"
      responses:
        "201":
          description: Created task id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnqueueTaskResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/tasks/{id}/cancel:
    post:
      tags: [tasks]
      operationId: cancelTask
      summary: Cancel a pending or running task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Cancelled
        "404":
          description: Not found / not cancellable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tasks/{id}/bump:
    post:
      tags: [tasks]
      operationId: bumpTask
      summary: Raise priority of a pending task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Bumped
        "404":
          description: Not pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/domains/{domain}/active:
    put:
      tags: [queue]
      operationId: setDomainActive
      summary: Set whether a domain is active (inactive domains do not run queue tasks; deactivating cancels pending and running tasks for that host)
      parameters:
        - $ref: "#/components/parameters/Domain"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetActiveRequest"
      responses:
        "204":
          description: Updated
        "400":
          description: Invalid domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/domains/{domain}/paused:
    put:
      tags: [queue]
      operationId: setDomainPaused
      summary: Soft-pause or resume a domain queue lane (ClaimNext blocked; interactive metadata prefetch still runs; pending stay queued; running continue; does not create a domains override row)
      parameters:
        - $ref: "#/components/parameters/Domain"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPausedRequest"
      responses:
        "204":
          description: Updated
        "400":
          description: Invalid domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/history:
    get:
      tags: [history]
      operationId: listHistory
      summary: Finished tasks (History)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [done, failed, cancelled]
          description: Filter by terminal task status
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: History items (finished tasks)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryItem"
  /api/notifications:
    get:
      tags: [history]
      operationId: listNotifications
      summary: In-app notification log
      parameters:
        - name: event
          in: query
          required: false
          schema:
            type: string
            enum: [cookie_invalid, rate_limited, ytdlp_failed, verify_failed, download_digest]
        - name: unread_only
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Notifications newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
...
  /api/notifications/{id}:
    get:
      tags: [history]
      operationId: getNotification
      summary: One notification
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/notifications/{id}/read:
    post:
      tags: [history]
      operationId: markNotificationRead
      summary: Mark one notification as read
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Notification after mark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/settings:
    get:
      tags: [settings]
      operationId: listSettings
      summary: List editable settings with help text
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SettingEntry"
    put:
      tags: [settings]
      operationId: updateSettings
      summary: Update one or more settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SettingEntry"
        "400":
          description: Unknown key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cookies:
    get:
      tags: [cookies]
      operationId: listCookies
      summary: List per-domain Netscape cookie jars
      responses:
        "200":
          description: Cookies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cookie"
    put:
      tags: [cookies]
      operationId: upsertCookie
      summary: Create or replace cookies for a domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCookieRequest"
      responses:
        "204":
          description: Saved
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cookies/{domain}:
    delete:
      tags: [cookies]
      operationId: deleteCookie
      summary: Delete cookies for a domain
      parameters:
        - $ref: "#/components/parameters/Domain"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    TaskId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Domain:
      name: domain
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine code (e.g. CookieInvalid, DownloadFailed, RemuxFailed, PackFailed)
        message:
          type: string
          description: Short user-facing text
        detail:
          type: string
          description: Optional operator/debug detail
...
    TaskStatus:
      type: string
      enum: [pending, running, done, failed, cancelled]
    Task:
      type: object
      required: [id, kind, status, domain, created_at]
      properties:
        id:
          type: integer
          format: int64
        kind:
          type: string
          description: scan, download, rescan_metadata, import
        status:
          $ref: "#/components/schemas/TaskStatus"
        domain:
          type: string
        series_id:
          type: integer
          format: int64
          nullable: true
        video_id:
          type: integer
          format: int64
          nullable: true
        message:
          type: string
        progress:
          type: number
          format: double
          nullable: true
        queue_position:
          type: integer
          nullable: true
          description: 1 = front of domain queue
        error_code:
          type: string
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
    EnqueueTaskRequest:
      type: object
      required: [kind, domain]
      properties:
        kind:
          type: string
        domain:
          type: string
        series_id:
          type: integer
          format: int64
        video_id:
          type: integer
          format: int64
        message:
          type: string
        priority:
          type: integer
    EnqueueTaskResponse:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
...
    HistoryItem:
      type: object
      required: [id, created_at, kind, status, message]
      properties:
        id:
          type: integer
          format: int64
          description: Task id
        created_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
        kind:
          type: string
        status:
          type: string
          enum: [done, failed, cancelled]
        message:
          type: string
        code:
          type: string
        detail:
          type: string
        series_id:
          type: integer
          format: int64
          nullable: true
        video_id:
          type: integer
          format: int64
          nullable: true
        domain:
          type: string
    Notification:
      type: object
      required: [id, created_at, event, title, body, external_ok, unread]
      properties:
        id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        event:
          type: string
          enum: [cookie_invalid, rate_limited, ytdlp_failed, verify_failed, download_digest]
        title:
          type: string
        body:
          type: string
        task_id:
          type: integer
          format: int64
          nullable: true
        external_ok:
          type: boolean
          description: True when any Apprise channel delivery succeeded
        read_at:
          type: string
          format: date-time
          nullable: true
        unread:
          type: boolean
          description: True for unread alert notifications
...
    SettingEntry:
      type: object
      required: [key, label, value, help]
      properties:
        key:
          type: string
          description: Stable settings key (snake_case)
        label:
          type: string
          description: Human-readable General Settings title
        value:
          type: string
        help:
          type: string
    UpdateSettingsRequest:
      type: object
      required: [values]
      properties:
        values:
          type: object
          additionalProperties:
            type: string
    Cookie:
      type: object
      required: [domain, content, updated_at]
      properties:
        domain:
          type: string
          description: Hostname (www. stripped) or `default` for the global fallback jar
        content:
          type: string
        updated_at:
          type: string
          format: date-time
    UpsertCookieRequest:
      type: object
      required: [domain, content]
      properties:
        domain:
          type: string
          description: Hostname or `default` for the global fallback jar
        content:
          type: string
...
    SetActiveRequest:
      type: object
      required: [active]
      properties:
        active:
          type: boolean
          description: Domain active flag; when false, queue tasks for this hostname do not run and pending+running tasks for the host are cancelled
    SetPausedRequest:
      type: object
      required: [paused]
      properties:
        paused:
          type: boolean
          description: Soft pause; when true, ClaimNext is blocked for this hostname. Interactive metadata prefetch still runs. Pending stay queued; running continue. Does not cancel tasks or create a domains override row.
...